round 6 digits of the float64 we recieve from rust in text_embeddings change to float64 because ruby by default produces float64, not sure if this makes a difference Solr's vector type is float32 but it didn't seem to have a problem with what we're sending from bibdata make the same change in orangelight
We investigated with @sandbergja the solr failure to index the record
when it has the 384 dimensions with the default float we return from Rust
We found that the problem happens only through ruby when we use the http post in pul solr json writer
We dont have this issue when we curl a json file that has not been generated
through the pul json writer-> see class and specifically method 'send_batch'

I tried indexing a smaller length of the vector dimensions for example 4 and it went
through the PulJsonWriter successfully. I'm not sure yet why.
Next step would be to compare the 384 dimension vector that is parsed in the
send_batch method vs the one that we get from the traject config and see how different are the text embeddings.

+++ embedding_service/src/embedding_client.rs
@@ -9,8 +9,7 @@ impl EmbeddingClient {
     pub fn new(url: String) -> Self {
         EmbeddingClient { url }
     }
-
-    pub fn get_embedding(&self, text: &str) -> Result<Vec<f32>, Box<dyn std::error::Error>> {
+    pub fn get_embedding(&self, text: &str) -> Result<Vec<f64>, Box<dyn std::error::Error>> {
         let mut url = Url::parse(&self.url)?;
         url.path_segments_mut()
             .map_err(|_| "Cannot modify URL path segments")?
@@ -41,14 +40,13 @@ impl EmbeddingClient {
             .map(|v| {
                 v.as_f64()
                     .ok_or("Expected embedding values to be floats")
-                    .map(|i| i as f32)
+                    .map(|i| i as f64)
             })
-            .collect::<Result<Vec<f32>, _>>()?;
-        // println!("Embedding: {:?}", embedding);
+            .collect::<Result<Vec<f64>, _>>()?;
         Ok(embedding)
     }
 }
-pub fn get_embedding(text: String) -> Result<Vec<f32>, magnus::Error> {
+pub fn get_embedding(text: String) -> Result<Vec<f64>, magnus::Error> {
     let base_url = std::env::var("EMBEDDING_SERVICE_URL")
         .unwrap_or_else(|_| "http://localhost:8000".to_string());
     let client = EmbeddingClient::new(base_url);
